requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b4
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.1
greenlet==1.1.3
idna==3.4
langdetect==1.0.9
magic-filter==1.0.9
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pydantic==1.10.2
Pygments==2.13.0
python-dotenv==0.21.0
python-md==1.0.2
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
typing_extensions==4.3.0
yarl==1.8.1
